fix(active-active): Fix start workflow with terminate if running policy for active-active domains (#7361)
<!-- Describe what has changed in this PR -->
**What changed?**
Update StartWorkflow and SignalWithStartWorkflow API to return a bad
request error, if the current workflow is active in a different cluster
with a different active cluster policy and the workflowID reuse policy
is terminate if running.

<!-- Tell your future self why have you made these changes -->
**Why?**
We don't support cross-region transactions, so we're unable to terminate
a workflow that's active in another cluster and start a new workflow in
the current cluster.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
simulation tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**