fix(active-active): cleanup clusters by region fields (#7357)
<!-- Describe what has changed in this PR -->
**What changed?**

this is a mechanical removal of the clustersByRegion and refactor
(except where otherwise remarked) as part of the cleanup to the
clusterAttributes schema.

It notably:
- Chooses to backwards-compat migrate the regional fields into the
cluster attributes schema
- Catches a few minor missed earlier updates 

Behaviour changes: 

This PR modifies a few Active/Active unit tests as the refactor removes
the old implmeentation of ClustersByRegion. This is a prerelease feature
that has been simplified and no actual known customer for it exists in
the wild. It's very likely unlikely any of these changes affect anyone
in any material fashion.

But for completeness:

- DomainConfig counter: Because the clustersByRegion concept is
implemented slightly differently to ClusterAttributes (in that the
DomainConfig counter is updated for Cluster Attributes on domain
failover, vs not in the ClustersByRegion Impl), this is therefore a
slight behavioural change here - the field DomainUpdate is incremented
on ClusterAttribute change and therefore the update is applied to the
domain in every case. I don't see any downside to this behaviour change
and it's clearer, but it's still worth flagging.
- FailoverVersion not applied to cluster attributes: Because these are
unrelated counters, cluster attributes no longer use or reference this
and act independendently on domain cluster attributes being updated.

**Testing** I manually tested it

---------

Signed-off-by: David Porter <david.porter@uber.com>