fix: correct handling of --ac flag for AA domains (#7362)
<!-- Describe what has changed in this PR -->
**What changed?**

This fixes a bug where the active-cluster flag is out of sync for AA
domains with the failover version, probably leading to some undefined,
or at least very confusing behaviour.

This fixes it to be in line with normal active/passive domains.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: David Porter <david.porter@uber.com>