fix: Another round of refactoring for disused fields (#7365)
<!-- Describe what has changed in this PR -->
**What changed?**
This does a bit more cleanup for some disused fields for the AA project.
These fields are not in use.

ClusterAttributes and CLi: 
- I'm not attempting to render the cluster-attributes in the normal
domain `desc` command since I'm not sure it's really a worthwhile thing
to render. It's visible already with the JSON output if details are
needed and honestly, working with the JSON is a lot more tractable.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>