feat: Handle Child Workflows with Ephemeral TaskLists (#7372)
Move the logic to overwrite the Activity TaskList to the decision
processing, where we check and overwrite most other decision attributes.
Share the logic between Child Workflows and Activities so they handle
this behavior the same.

This removes the responsibility from the client for specifying the
correct TaskListKind for child workflows of a Workflow started with an
Ephemeral TaskList.

This additionally introduces a minor change in how we process Activity
TaskLists, aligning it with Child Workflows. Child Workflows don't
require a TaskList and default to the Workflow's TaskList. Activities
require a TaskList and the server fails the decision if it's not
present.

Within each client we currently have logic that makes the activity
TaskList default to the Workflow TaskList. Adding a redundant copy of
this behavior to the server is low risk.

<!-- Describe what has changed in this PR -->
**What changed?**
- Child Workflows started by a Workflow with an Ephemeral TaskList will
similarly run on an Ephemeral TaskList

<!-- Tell your future self why have you made these changes -->
**Why?**
- Propagating the Ephemeral nature

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
- Unit and integration tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
- Low. This adds no additional validation, and we haven't yet enabled or
announced this feature. Even if it was actively used, the TaskListKind
is determined when the TaskListManager is created. Treating an Ephemeral
TaskList as normal will cause no issues and is very low risk, as it's
effectively disabling the feature.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**