feat: adding failover history endpoint (#7381)
<!-- Describe what has changed in this PR -->
**What changed?**

This scaffolds a new endpoint for domain failover history, a list/read
only endpoint which is intended to enumerate the recent failover related
events that affect a domain. It's intended to be used in the UI for
letting domain-owners know what's occurred. It's going to be used in the
Active/Active feature, but is broadly applicable to all failover events
including the normal active/passive failovers and is going to replace
the hacky domain_data field.

**Why?**

Because we were stuffing some failover information in the domain data
for a while, but that was frankly.... not very good. It wasn't really a
serious attempt at an API, nor did it hold any large amount of history
and, perhaps most imporantly, wouldn't have scaled with any additional
data.

Domain-level auditing is a more general problem and the features that
@c-warren was working on are intended to capture all these and put them
in the domain-events table, of which failover is a special case (getting
it's own endpoint).

**How did you test it?**
- [x] Unit tests. This feature's not implemented, scaffolded only.

**Detailed Description**

As above

**Impact Analysis**

- **Backward Compatibility**: 
This is backwards compatible, it's just adding a new endpdoint
- **Forward Compatibility**: 
It should be forwards compatible

**Testing Plan**
- **Unit Tests**: to add
- **Persistence Tests**: to add
- **Integration Tests**: to add
- **Compatibility Tests**:  Not needed right now

**Rollout Plan**

This will be rolled out once it's implemented and the UI will consume it
optionally.

---------

Signed-off-by: David Porter <david.porter@uber.com>