feat: Adds mappers, basic CLI, testdata for ListDomainFailover endpoint (#7389)
<!-- Describe what has changed in this PR -->
**What changed?**

The endpoint for the domain metadata / failover storage information
needs some more scaffolding. To test it and unblock the UI folks, I'm
adding a bit of temporary data there while it's in flux and being worked
on.

This change:
- Adds mappers for thrift/proto for this endpoint
- Adds a basic CLi for interacting with the endpoint 
- Adds some testdata

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

- [X] Tested running locally with the CLI
- [X] some, (though maybe not enough) unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

This endpoint isn't in use yet, so very low risk, additive changes only
here

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>