chore: adds a deepcopy method domain-update handler refactor (#7390)
<!-- Describe what has changed in this PR -->
**What changed?**

The domain UpdateHandler needs a refactor. Part of what makes it very
confusing and hard to follow is the way in which it mutates the current
state struct to create a configuration the domain should be in. It's
hard, later in the function, to then go back and lookup the current
state.

As a partial mitigation, I'm going to deepcopy the state such that I can
use it in a subequent refactor

There's no deepcopy function on the GetDomainResponse struct right now.
Arguably it's a terrible, ugly mess of additional code and... I don't
love it, but my hope is that we can codegen this and other structs soon
so as to not maintain these manually copied structs, since I don't think
this is scalable.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit tests, no expected behaviour change.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>