feat: [shard-distributor] remove deprecated service instantiation (#7387)
<!-- Describe what has changed in this PR -->
**What changed?**
We are cleaning up the deprecated way of instantiating the service. 

<!-- Tell your future self why have you made these changes -->
**Why?**
Fx injection should be use, to reduce tech debt we should clean up since
we are not keeping this updated with new dependencies.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
no failing unit tests, the local instance of the service can start

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
the service is in development mode, there are no dependencies on this
 
<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: edigregorio <edigregorio@uber.com>