fix: add nil pointer check for active cluster selection policy mappers (#7398)
<!-- Describe what has changed in this PR -->
**What changed?**
- add nil pointer check for active cluster selection policy mappers

<!-- Tell your future self why have you made these changes -->
**Why?**
We want to remove those fields from ActiveClusterSelectionPolicy.
However, the change is not safe to rollback because of the nil-pointer
dereference of the new data that doesn't have those fields serialized
into thrift. This change is to create a commit that's safe to be rolled
back.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**