fix: Failover-history mapping bug for thrift (#7407)
<!-- Describe what has changed in this PR -->
**What changed?**

Fixes an embarrasing thrift mapper bug that I left when merging. The
more interesting question is why the test didn't pick it up. The reason
was that it was manually constructing the failover event (which was
nil), thus not actually generating data for it. I don't remember why
that was present, but it doesn't seem valid.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

- Fixed / via unit tests 
- Manually 

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: David Porter <david.porter@uber.com>