fix: [shard-distributor-canary] Fix canary startup (#7411)
<!-- Describe what has changed in this PR -->
**What changed?**
Pass the names of namespace as parameters


<!-- Tell your future self why have you made these changes -->
**Why?**
For:
- shard-distributor-canary
- shard-distributor-canary-ephemeral
- test-external-assignment
we want to test namespaces based on environment and region, so we need
to have it passed as parameter and not statically defined in the module
definition

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
tested locally and in the staging environment

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
canary is not running, after this it can be started again and generate
load on etcd

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>