fix: upgrade go toolchain and mockery (#7414)
<!-- Describe what has changed in this PR -->
**What changed?**

This upgrades some internal tooling and the mockery library which is
blocking upgrades for other things (nilaway amoung other things)

<!-- Tell your future self why have you made these changes -->
**Why?**

I think we should probably finish removing mockery (I have a mostly-done
PR for this), but in the mean-time this is a straightforward upgrade and
unblocker.


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>