feat(active-active): Introduce a feature flag to enable active cluster selection policy (#7417)
<!-- Describe what has changed in this PR -->
**What changed?**
Introduce a feature flag to enable active cluster selection policy

<!-- Tell your future self why have you made these changes -->
**Why?**
With the feature flag, we can have new workflows in an active-active
domain to behave as in an active-passive domain, so that in case some
bug occurs we have a way to fallback safely.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**