feat: adds nilaway (nonblocking) for mappers (#7413)
<!-- Describe what has changed in this PR -->
**What changed?**

This adds the `nilaway` check to the repo's mappers to enforce a degree
of safety. The intent is that the mapper-code should be more defensive
and dereference pointers defensively both for general safety as well as
forwards-compatibility.

This: 
- Currently is nonblocking, it won't fail the lint step
- There are a few dozen cases it's currently matching which are a
problem
- I expect to clean up mappers in a future PR and enforce mapping

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

none expected, tooling only 

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>