fix: Handle duplicate child workflow events (#7400)
We should certainly explore a more generic or generally better solution
for this (maybe comparing HistoryEvent attributes?), but for now I think
it's reasonable to repeat the fix we did for activities. We still don't
know the root cause of this.

<!-- Describe what has changed in this PR -->
**What changed?**
- Handle duplicate buffered Child Workflow Events

<!-- Tell your future self why have you made these changes -->
**Why?**
- We recently observed a scenario where these were duplicated, similar
to activities.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
- Unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**