feat: [shard-distributor] rely on dynamic config for migration mode (#7391)
<!-- Describe what has changed in this PR -->
**What changed?**
Move the logic for the migration mode to use dynamic config instead of
static config. Next step is to clean up the static config once the
behavior is tested with multiple configuration changes.

<!-- Tell your future self why have you made these changes -->
**Why?**
migration should be done incrementally and this is easier to control
with config rollout and not code rollout.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests and local development.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
shard-distributor-canary service cannot operate correctly if
shard-distributor do not use the right configuration

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>