feat: scaffolding domain audit manager (#7422)
<!-- Describe what has changed in this PR -->
**What changed?**

This adds a domain-audit functionality and implements the subset of
events on it for failovers in the ListFailoverHistory endpoint for
reading this data. The intent of this feature is that for any
significant create, write, failover or other important domain operation,
to put the events into an audit log for tracking. This supercedes the
5-most-recent-failovers domain-data blob that was previously created
quickly.

This PR only expose that read information for failovers, however, it
persists all write operations to the audit log, with the expecation that
a more general-purpose audit read endpoint may be added in the future.

This is a scaffolding PR and it's probably missing a little in the way
of features, among them:

- [ ] TTL for data cleanup
- [ ] Probably some test coverage

These are blockers before landing and they'll be included in the PR

Testing:

- [X] Manual testing
- [X] Unit testing

Signed-off-by: David Porter <david.porter@uber.com>